backend/notice_project/requirements.txt
APScheduler==3.8.0
asgiref==3.4.1
astroid==2.8.0
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cfgv==3.3.1
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.0.1
distlib==0.3.3
Django==3.2.6
django-cors-headers==3.8.0
django-environ==0.4.5
djangorestframework==3.12.4
docformatter==1.4
drf-yasg==1.20.0
Faker==9.2.0
filelock==3.3.0
flake8==3.9.2
gunicorn==20.1.0
identify==2.3.0
idna==3.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
nodeenv==1.6.0
openapi-codec==1.3.2
packaging==21.0
platformdirs==2.4.0
pre-commit==2.15.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.11.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.1
tzlocal==2.1
untokenize==0.1.1
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.8.1
wrapt==1.12.1
